Refactoring utility methods.
